refactor(verify-git-email): re-name test file to reflect source structure
