+++ tests/verify_git_email_test.py
@@ -1,31 +1,6 @@
-import os
-import pathlib
-import subprocess
-
 import pytest
 
 from pep_pre_commit_hooks import verify_git_email
-
-
-@pytest.fixture()
-def _ch_tempdir(tmp_path):
-    """Temporarily change the working directory to a temporary directory."""
-    cwd = pathlib.Path.cwd()
-    try:
-        os.chdir(tmp_path)
-        yield
-    finally:
-        os.chdir(cwd)
-
-
-@pytest.fixture()
-def _git_init():
-    subprocess.check_call(("git", "init"))
-
-
-@pytest.fixture()
-def _git_config_icloud_email():
-    subprocess.check_call(("git", "config", "user.email", "test@icloud.com"))
 
 
 @pytest.mark.usefixtures("_ch_tempdir", "_git_init", "_git_config_icloud_email")
@@ -40,3 +15,12 @@ def test_failed_verify_domain():
         match="but an email address matching ",
     ):
         verify_git_email.verify_git_email("hotmail.com")
+
+
+@pytest.fixture(params=(("--domain", "@live.com"),))
+def domain(request):
+    return request.param
+
+
+def test__parse_args(domain):
+    assert verify_git_email._parse_args(domain).domain == domain[1]
